% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_blockchain.R
\name{blockchain}
\alias{blockchain}
\title{R6 Class representing a blockchain}
\description{
R6 Class representing a blockchain

R6 Class representing a blockchain
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{blocks}}{List of blocks.}

\item{\code{transaction_queue}}{transaction queue
New transaction}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-blockchain-new}{\code{blockchain$new()}}
\item \href{#method-blockchain-greet}{\code{blockchain$greet()}}
\item \href{#method-blockchain-hash}{\code{blockchain$hash()}}
\item \href{#method-blockchain-validation}{\code{blockchain$validation()}}
\item \href{#method-blockchain-genenis_block}{\code{blockchain$genenis_block()}}
\item \href{#method-blockchain-new_block}{\code{blockchain$new_block()}}
\item \href{#method-blockchain-chain_validation}{\code{blockchain$chain_validation()}}
\item \href{#method-blockchain-add_block}{\code{blockchain$add_block()}}
\item \href{#method-blockchain-get_address}{\code{blockchain$get_address()}}
\item \href{#method-blockchain-check_balance}{\code{blockchain$check_balance()}}
\item \href{#method-blockchain-check_transaction}{\code{blockchain$check_transaction()}}
\item \href{#method-blockchain-new_transaction}{\code{blockchain$new_transaction()}}
\item \href{#method-blockchain-chain_data}{\code{blockchain$chain_data()}}
\item \href{#method-blockchain-balances}{\code{blockchain$balances()}}
\item \href{#method-blockchain-print_block}{\code{blockchain$print_block()}}
\item \href{#method-blockchain-chain_info}{\code{blockchain$chain_info()}}
\item \href{#method-blockchain-print_chain}{\code{blockchain$print_chain()}}
\item \href{#method-blockchain-generate_key}{\code{blockchain$generate_key()}}
\item \href{#method-blockchain-clone}{\code{blockchain$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-new"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-new}{}}}
\subsection{Method \code{new()}}{
Create a new blockchain object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Extra parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{blockchain} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-greet"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-greet}{}}}
\subsection{Method \code{greet()}}{
Initialization messag
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$greet()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A initialization message.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-hash"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-hash}{}}}
\subsection{Method \code{hash()}}{
Hash function sha256
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$hash(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Object to hash}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A hashed char
Validation function searching a nonce for a hash beggining with 000
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-validation"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-validation}{}}}
\subsection{Method \code{validation()}}{
Validate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$validation(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Object to validate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Object validate and hashed
Genesis block generation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-genenis_block"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-genenis_block}{}}}
\subsection{Method \code{genenis_block()}}{
Generate first block
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$genenis_block()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
Generate a block
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-new_block"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-new_block}{}}}
\subsection{Method \code{new_block()}}{
Generate a block
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$new_block()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
Function to check the chain validation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-chain_validation"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-chain_validation}{}}}
\subsection{Method \code{chain_validation()}}{
Generate a block
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$chain_validation()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
Add the block to chain
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-add_block"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-add_block}{}}}
\subsection{Method \code{add_block()}}{
Add a block
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$add_block()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
Extract public key
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-get_address"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-get_address}{}}}
\subsection{Method \code{get_address()}}{
Extract public key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$get_address(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{key}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Encoded base64 puk key
Check bala
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-check_balance"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-check_balance}{}}}
\subsection{Method \code{check_balance()}}{
Extract public key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$check_balance(address)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{address}}{key}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Encoded base64 puk key
Check transaction
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-check_transaction"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-check_transaction}{}}}
\subsection{Method \code{check_transaction()}}{
Check transaction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$check_transaction(transaction)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{transaction}}{transaction}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Encoded base64 puk key
New transaction
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-new_transaction"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-new_transaction}{}}}
\subsection{Method \code{new_transaction()}}{
New transaction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$new_transaction(path_private, receiver, amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path_private}}{private key}

\item{\code{receiver}}{pub key recever}

\item{\code{amount}}{amount}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-chain_data"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-chain_data}{}}}
\subsection{Method \code{chain_data()}}{
New transaction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$chain_data()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path_private}}{private key}

\item{\code{receiver}}{pub key recever}

\item{\code{amount}}{amount}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame of transactions
account balances
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-balances"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-balances}{}}}
\subsection{Method \code{balances()}}{
account balances
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$balances()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
account balances
Print a block
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-print_block"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-print_block}{}}}
\subsection{Method \code{print_block()}}{
Print a block
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$print_block(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{block}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Print a block
Number of blocks
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-chain_info"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-chain_info}{}}}
\subsection{Method \code{chain_info()}}{
Number of blocks
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$chain_info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Number of blocks
Print chain
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-print_chain"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-print_chain}{}}}
\subsection{Method \code{print_chain()}}{
Print chain
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$print_chain()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Print chain
Generate keys
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-generate_key"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-generate_key}{}}}
\subsection{Method \code{generate_key()}}{
Generate a keys
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$generate_key(private, public)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{private}}{private path}

\item{\code{public}}{public public path}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Generate key
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blockchain-clone"></a>}}
\if{latex}{\out{\hypertarget{method-blockchain-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blockchain$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
